Enhance "smtp" component filter (#17)
* Enhance "smtp" component filter

fixes #16

This filter now mainly consists of a big GREEDYDATA pattern that will
match on all information postfix throws at us.

But if the message contains extra (structured) information it will parse
that as well. This will help to work around the fact that sometimes only
the detail information is the whole message and sometimes it's
encapsulated in () with more information

* Re-Add "grokked" tag. Open issue for detail

I merge without review because this is marked as "not production ready" anyway.